Extract RegexExpressionHelper
