+++ conf/config.neon
@@ -1492,6 +1492,9 @@ services:
 		class: PHPStan\Type\Php\RegexArrayShapeMatcher
 
 	-
+		class: PHPStan\Type\Php\RegexExpressionHelper
+
+	-
 		class: PHPStan\Type\Php\ReflectionClassConstructorThrowTypeExtension
 		tags:
 			- phpstan.dynamicStaticMethodThrowTypeExtension
